ts-cli: rename proxy -> env-proxy, move test next to source
Committed-By-Agent: claude
